+++ src/components/month/calendarOpenDayEvents.component.ts
@@ -1,7 +1,6 @@
 import {
   Component,
   Input,
-  ChangeDetectionStrategy,
   trigger,
   style,
   transition,
@@ -13,13 +12,18 @@ import { CalendarEvent } from 'calendar-utils';
 
 @Component({
   selector: 'mwl-calendar-open-day-events',
-  changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
     <div class="cal-open-day-events" [@collapse] *ngIf="isOpen">
       <div
         *ngFor="let event of events"
-        [ngClass]="event?.cssClass">
-        <span class="cal-event" [style.backgroundColor]="event.color.primary"></span>
+        [ngClass]="event?.cssClass"
+        mwlDraggable
+        [dropData]="{event: event}"
+        [dragAxis]="{x: event.draggable, y: event.draggable}">
+        <span
+          class="cal-event"
+          [style.backgroundColor]="event.color.primary">
+        </span>
         <mwl-calendar-event-title
           [event]="event"
           view="month"