+++ src/interfaces/calendarEventTimesChangedEvent.interface.ts
@@ -3,5 +3,5 @@ import { CalendarEvent } from 'calendar-utils';
 export interface CalendarEventTimesChangedEvent {
   event: CalendarEvent;
   newStart: Date;
-  newEnd: Date;
+  newEnd?: Date;
 }